api/requirements.txt
django
djangorestframework
psycopg2-binary
factory_boy
pandas
ipython
jupyter
django-extensions
xlrd
django-pdb
ipdb
djangorestframework-gis
graphql-core>=2.0.dev
graphene-django
django-graphql-geojson
django-filter>=2